Moved these things out of the pages directory because they are not pages.
