+++ imports/ui/pages/contact-us.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import Header from "../../ui/pages/header.jsx";
-import Footer from "../../ui/pages/footer.jsx";
+import Header from "/imports/ui/components/header.jsx";
+import Footer from "/imports/ui/components/footer.jsx";
 import { Meteor } from "meteor/meteor";
 
 /* A page where visitors can contact Vize for buisness inqueries.
