+++ imports/ui/showjobs.jsx
@@ -3,7 +3,7 @@ import JobPosting from "../ui/components/jobPosting.jsx";
 import { JobAds } from "../api/data/jobads.js";
 import { withTracker } from "meteor/react-meteor-data";
 import ShowJobComponent from "../ui/components/showJobComponent.jsx";
-import Header from "./pages/header.jsx";
+import Header from "/imports/ui/components/header.jsx";
 import i18n from "meteor/universe:i18n";
 
 const T = i18n.createComponent();
@@ -11,7 +11,11 @@ const T = i18n.createComponent();
 class ShowJobs extends React.Component {
 	render() {
 		if (!this.props.isReady) {
-			return <h2><T>common.jobsearch.loading</T></h2>;
+			return (
+				<h2>
+					<T>common.jobsearch.loading</T>
+				</h2>
+			);
 		}
 
 		const RenderedItems = this.props.jobads.map(function(jobad) {
@@ -19,7 +23,11 @@ class ShowJobs extends React.Component {
 		});
 		let message;
 		if (RenderedItems.length < 1) {
-			message = <h2><T>common.jobsearch.nojobs</T></h2>;
+			message = (
+				<h2>
+					<T>common.jobsearch.nojobs</T>
+				</h2>
+			);
 		} else {
 			message = "";
 		}
@@ -38,7 +46,8 @@ class ShowJobs extends React.Component {
 							<ul className="rslides" id="slider3">
 								<li>
 									<h2>
-										{this.props.numberofjobs} <T>common.jobsearch.jobsAvailable</T>
+										{this.props.numberofjobs}{" "}
+										<T>common.jobsearch.jobsAvailable</T>
 									</h2>
 									{message}
 									{RenderedItems}
